Update Unicode snippets and mark ".t" files as Perl
The Turing package I made maps the ".t" extension to the Turing grammar,
overriding the Perl package's association. Ordinarily I'd just uninstall
the damn offending package, but in this case, it's one of mine. Lel.

Unfortunate caveat: this screws with the icon assigned by File-Icons to
Perl unit-tests (replacing the distinctive beaker icon with the default
for Perl files). Should do something about this for File-Icons v2.2.0.