Globalise Disposable classes and disable atomic tabs
I never use spaces to indent. So if there's a file with soft tabs opened
in Atom, it'll be because it's not my work. Which probably means it uses
a different tabstop width to the one I prefer.

This makes it harder to edit soft-tabbed code set to a different tabstop
length, because Atom or one of its packages resets a buffer's tab-length
after it's been saved.