Version-lock and downgrade Minimap to v4.29.9
Rather than report the breakage that recent releases have inflicted upon
me, I'd rather stick to the same version I've been using for years. That
same idiot I cracked the shits with and blocked for file-icons/atom-fs#8
is to "thank" for recent releases, so filing an issue might be awkward.