Add command to format SVG paths exported by Illustrator
This currently only formats the `d` attributes of `<path/>` elements. It
produces output that's consistent and size-optimal, albeit unreadable by
humans. This is suitable for semi-readable SVGs where structure, styling
and/or translations are designed to be editable using a text-editor, but
individual drawing coordinates remain opaque. For example:

    <?xml version="1.0" encoding="utf-8"?>
    <svg viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg">
        <!-- Basic chevron with human-readable components -->
        <path id="simple-chevron" d="
            M 20   5
            l  5  10
            l -5  10
            l 10 -10
            z
        "/>

        <!-- Fancy-looking but opaque chevron -->
        <path id="fancy-chevron" d="
            M4,5c0,3.8,3.2,7.4,5.1,10C7.1,17.7,4,21.3,4,25c2.3-4,
            5.1-6.9,8.3-8.8c0.8-0.5,0.7-1.7,0-2.2C9.1,11.9,6.1,8.5,
            4,5zM6.2,15c0,0.7-0.5,1.2-1.2,1.2S3.8,15.7,3.8,15
            s0.5-1.2,1.2-1.2S6.2,14.3,6.2,15z
        "/>
    </svg>

The former is amenable to multi-cursor editing, and is understandable by
anybody with a rudimentary understanding of SVG code. The latter is much
less readable, but still benefits from having visibly-distinct shapes on
separate lines. Running it through `user:format-svg` produces:

    <path id="fancy-chevron" d="
        M4 5c0 3.8 3.2 7.4 5.1 10C7.1 17.7 4 21.3 4 25c2.3-4 5.1-6.9…
        M6.2 15c0 0.7-0.5 1.2-1.2 1.2S3.8 15.7 3.8 15s0.5-1.2 1.2-1.2…
    "/>

Note that the command currently requires each `<path>` element to reside
entirely within an editor's selection in order for its coordinates to be
properly formatted. This may or may not be fixed in future, depending on
whether I can be arsed.