Add the most niche editor-command I've penned to date
I'm trying to isolate historical words that appear in the subtitles of a
1930s movie, and having this weird-looking number prepended to a list of
each unique word extracted from an SRT file would enable me to sort them
numerically and locate the lesser-used terms.

Like I said: unbelievably specific, but I wasted a groggy all-nighter on
this crap, and I'm damned if I'm not saving the fruits of my labour.