Add function for stripping ID3 tags from WAV files
After ripping a CD losslessly using OpenBSD's cdio(1) command, Picard is
used to generate Acoustid fingerprints for each track, provided an entry
for the release exists on MusicBrainz.org. These are embedded in the WAV
files as trailing ID3 metadata, invalidating the SHA256 checksums which,
ideally, should be computed using the CD's raw audio data.

As hyper-specific as this scenario is, it's happened enough times for it
to warrant writing this function.