Add modeline support for OpenStep property lists
Neither Vim nor Emacs support plist(5) files, irrespective of which file
format is used. I'm assuming that an author of a plugin or package would
choose `plist` as the syntax/major-mode name, however.

Also, fix the icon shown for files named `.meta`, which I've tentatively
chosen as the filename to dump useful metadata to that would normally be
lost upon copying to an FFS2 filesystem (i.e., my portable hard-drive).