Add manually-ported version of GitHub's blob-anchor module
This component is loaded as part of GitHub's front-end to highlight line
(and/or column) ranges extracted from the blob permalink's anchor text:

    #L5         := Highlights line 5
    #L2-L4      := Highlights lines 2-4 (inclusive)
    #L5C2-L8C5  := Highlights line 5 (starting at column 2),
                   through to line 8 (ending at column 5).

The original code, written in TypeScript, doesn't appear to be public; a
copy was lifted from the source-mapped original version of an obfuscated
and bundled JavaScript file.

Currently, the functions defined in this file aren't used by the rest of
my Atom setup; the expectation is to use a more concise, space-efficient
implementation written for Alhadis/language-rfc@7c2d412 (lines 342-350).